wip : get all logistic request
